update security model

+++ src/main/java/fr/istic/web/rest/CommentsResource.java
@@ -2,8 +2,13 @@ package fr.istic.web.rest;
 
 import static javax.ws.rs.core.UriBuilder.fromPath;
 
+import fr.istic.domain.Authority;
+import fr.istic.domain.Comments;
+import fr.istic.domain.Student;
+import fr.istic.domain.User;
 import fr.istic.security.AuthoritiesConstants;
 import fr.istic.service.CommentsService;
+import fr.istic.web.rest.errors.AccountResourceException;
 import fr.istic.web.rest.errors.BadRequestAlertException;
 import fr.istic.web.util.HeaderUtil;
 import fr.istic.web.util.ResponseUtil;
@@ -13,6 +18,8 @@ import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import fr.istic.service.Paged;
+import fr.istic.service.SecurityService;
+import fr.istic.service.StudentService;
 import fr.istic.web.rest.vm.PageRequestVM;
 import fr.istic.web.rest.vm.SortRequestVM;
 import fr.istic.web.util.PaginationUtil;
@@ -44,6 +51,10 @@ public class CommentsResource {
 
     @Inject
     CommentsService commentsService;
+
+    @Inject
+    SecurityService securityService;
+
     /**
      * {@code POST  /comments} : Create a new comments.
      *
@@ -73,11 +84,15 @@ public class CommentsResource {
      */
     @PUT
     @RolesAllowed({AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN})
-    public Response updateComments(CommentsDTO commentsDTO) {
+    public Response updateComments(CommentsDTO commentsDTO, @Context SecurityContext ctx) {
         log.debug("REST request to update Comments : {}", commentsDTO);
         if (commentsDTO.id == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
+        if (!securityService.canAccess(ctx, commentsDTO.id, Comments.class  )){
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        }
+
         var result = commentsService.persistOrUpdate(commentsDTO);
         var response = Response.ok().entity(result);
         HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, commentsDTO.id.toString()).forEach(response::header);
@@ -93,8 +108,12 @@ public class CommentsResource {
     @DELETE
     @RolesAllowed({AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN})
     @Path("/{id}")
-    public Response deleteComments(@PathParam("id") Long id) {
+    public Response deleteComments(@PathParam("id") Long id, @Context SecurityContext ctx) {
         log.debug("REST request to delete Comments : {}", id);
+        if (!securityService.canAccess(ctx, id, Comments.class  )){
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        };
+
         commentsService.delete(id);
         var response = Response.noContent();
         HeaderUtil.createEntityDeletionAlert(applicationName, true, ENTITY_NAME, id.toString()).forEach(response::header);
@@ -108,7 +127,7 @@ public class CommentsResource {
      * @return the {@link Response} with status {@code 200 (OK)} and the list of comments in body.
      */
     @GET
-    public Response getAllComments(@BeanParam PageRequestVM pageRequest, @BeanParam SortRequestVM sortRequest, @Context UriInfo uriInfo) {
+    public Response getAllComments(@BeanParam PageRequestVM pageRequest, @BeanParam SortRequestVM sortRequest, @Context UriInfo uriInfo, @Context SecurityContext ctx) {
         log.debug("REST request to get a page of Comments");
         var page = pageRequest.toPage();
         var sort = sortRequest.toSort();
@@ -119,9 +138,23 @@ public class CommentsResource {
             result = commentsService.findCommentsbyZonegeneratedid(page,  zonegeneratedid.get(0));
         }
         else {
-
-            result =commentsService.findAll(page);
+            var userLogin = Optional
+            .ofNullable(ctx.getUserPrincipal().getName());
+        if (!userLogin.isPresent()){
+            throw new AccountResourceException("Current user login not found");
         }
+        var user = User.findOneByLogin(userLogin.get());
+        if (!user.isPresent()) {
+            throw new AccountResourceException("User could not be found");
+        }
+            else if (user.get().authorities.size() >= 1 && user.get().authorities.stream().anyMatch(e1-> e1.equals(new Authority("ROLE_ADMIN")))){
+                result =commentsService.findAll(page);
+
+            } else {
+                return Response.status(403, "Current user cannot access to this ressource").build();
+            }
+
+         }
         var response = Response.ok().entity(result.content);
         response = PaginationUtil.withPaginationInfo(response, uriInfo, result);
         return response.build();
@@ -137,8 +170,11 @@ public class CommentsResource {
     @GET
     @Path("/{id}")
 
-    public Response getComments(@PathParam("id") Long id) {
+    public Response getComments(@PathParam("id") Long id, @Context SecurityContext ctx) {
         log.debug("REST request to get Comments : {}", id);
+        if (!securityService.canAccess(ctx, id, Comments.class  )){
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        };
         Optional<CommentsDTO> commentsDTO = commentsService.findOne(id);
         return ResponseUtil.wrapOrNotFound(commentsDTO);
     }
