+++ src/main/java/fr/istic/web/rest/StudentResponseResource.java
@@ -3,6 +3,7 @@ package fr.istic.web.rest;
 import static javax.ws.rs.core.UriBuilder.fromPath;
 
 import fr.istic.service.StudentResponseService;
+import fr.istic.web.rest.errors.AccountResourceException;
 import fr.istic.web.rest.errors.BadRequestAlertException;
 import fr.istic.web.util.HeaderUtil;
 import fr.istic.web.util.ResponseUtil;
@@ -12,8 +13,12 @@ import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import fr.istic.domain.Authority;
+import fr.istic.domain.StudentResponse;
+import fr.istic.domain.User;
 import fr.istic.security.AuthoritiesConstants;
 import fr.istic.service.Paged;
+import fr.istic.service.SecurityService;
 import fr.istic.web.rest.vm.PageRequestVM;
 import fr.istic.web.rest.vm.SortRequestVM;
 import fr.istic.web.util.PaginationUtil;
@@ -45,6 +50,10 @@ public class StudentResponseResource {
 
     @Inject
     StudentResponseService studentResponseService;
+
+    @Inject
+    SecurityService securityService;
+
     /**
      * {@code POST  /student-responses} : Create a new studentResponse.
      *
@@ -74,11 +83,15 @@ public class StudentResponseResource {
      */
     @PUT
     @RolesAllowed({AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN})
-    public Response updateStudentResponse(StudentResponseDTO studentResponseDTO) {
+    public Response updateStudentResponse(StudentResponseDTO studentResponseDTO, @Context SecurityContext ctx) {
         log.debug("REST request to update StudentResponse : {}", studentResponseDTO);
         if (studentResponseDTO.id == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
+        if (!securityService.canAccess(ctx, studentResponseDTO.id, StudentResponse.class  )){
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        }
+
         var result = studentResponseService.persistOrUpdate(studentResponseDTO);
         var response = Response.ok().entity(result);
         HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, studentResponseDTO.id.toString()).forEach(response::header);
@@ -94,8 +107,12 @@ public class StudentResponseResource {
     @DELETE
     @RolesAllowed({AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN})
     @Path("/{id}")
-    public Response deleteStudentResponse(@PathParam("id") Long id) {
+    public Response deleteStudentResponse(@PathParam("id") Long id, @Context SecurityContext ctx) {
         log.debug("REST request to delete StudentResponse : {}", id);
+        if (!securityService.canAccess(ctx, id, StudentResponse.class  )){
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        }
+
         studentResponseService.delete(id);
         var response = Response.noContent();
         HeaderUtil.createEntityDeletionAlert(applicationName, true, ENTITY_NAME, id.toString()).forEach(response::header);
@@ -109,7 +126,7 @@ public class StudentResponseResource {
      * @return the {@link Response} with status {@code 200 (OK)} and the list of studentResponses in body.
      */
     @GET
-    public Response getAllStudentResponses(@BeanParam PageRequestVM pageRequest, @BeanParam SortRequestVM sortRequest, @Context UriInfo uriInfo) {
+    public Response getAllStudentResponses(@BeanParam PageRequestVM pageRequest, @BeanParam SortRequestVM sortRequest, @Context UriInfo uriInfo, @Context SecurityContext ctx) {
         log.debug("REST request to get a page of StudentResponses");
         var page = pageRequest.toPage();
         var sort = sortRequest.toSort();
@@ -122,11 +139,24 @@ public class StudentResponseResource {
             Long.parseLong("" + questionId.get(0))
             );
         } else{
-            result =studentResponseService.findAll(page);
+
+            var userLogin = Optional
+            .ofNullable(ctx.getUserPrincipal().getName());
+        if (!userLogin.isPresent()){
+            throw new AccountResourceException("Current user login not found");
         }
+        var user = User.findOneByLogin(userLogin.get());
+        if (!user.isPresent()) {
+            throw new AccountResourceException("User could not be found");
+        }
+            else if (user.get().authorities.size() >= 1 && user.get().authorities.stream().anyMatch(e1-> e1.equals(new Authority("ROLE_ADMIN")))){
+                result =studentResponseService.findAll(page);
 
+            } else {
+                return Response.status(403, "Current user cannot access to this ressource").build();
+            }
 
-
+        }
         var response = Response.ok().entity(result.content);
         response = PaginationUtil.withPaginationInfo(response, uriInfo, result);
         return response.build();
@@ -141,8 +171,13 @@ public class StudentResponseResource {
      */
     @GET
     @Path("/{id}")
+    @RolesAllowed({AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN})
 
-    public Response getStudentResponse(@PathParam("id") Long id) {
+    public Response getStudentResponse(@PathParam("id") Long id, @Context SecurityContext ctx) {
+        if (!securityService.canAccess(ctx, id, StudentResponse.class  )){
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        }
+
         log.debug("REST request to get StudentResponse : {}", id);
         Optional<StudentResponseDTO> studentResponseDTO = studentResponseService.findOne(id);
         return ResponseUtil.wrapOrNotFound(studentResponseDTO);
