+++ src/main/java/fr/istic/web/rest/TextCommentResource.java
@@ -3,6 +3,7 @@ package fr.istic.web.rest;
 import static javax.ws.rs.core.UriBuilder.fromPath;
 
 import fr.istic.service.TextCommentService;
+import fr.istic.web.rest.errors.AccountResourceException;
 import fr.istic.web.rest.errors.BadRequestAlertException;
 import fr.istic.web.util.HeaderUtil;
 import fr.istic.web.util.ResponseUtil;
@@ -12,8 +13,12 @@ import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import fr.istic.domain.Authority;
+import fr.istic.domain.TextComment;
+import fr.istic.domain.User;
 import fr.istic.security.AuthoritiesConstants;
 import fr.istic.service.Paged;
+import fr.istic.service.SecurityService;
 import fr.istic.web.rest.vm.PageRequestVM;
 import fr.istic.web.rest.vm.SortRequestVM;
 import fr.istic.web.util.PaginationUtil;
@@ -42,6 +47,9 @@ public class TextCommentResource {
     @ConfigProperty(name = "application.name")
     String applicationName;
 
+
+    @Inject
+    SecurityService securityService;
 
     @Inject
     TextCommentService textCommentService;
@@ -74,10 +82,13 @@ public class TextCommentResource {
      */
     @PUT
     @RolesAllowed({AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN})
-    public Response updateTextComment(TextCommentDTO textCommentDTO) {
+    public Response updateTextComment(TextCommentDTO textCommentDTO, @Context SecurityContext ctx) {
         log.debug("REST request to update TextComment : {}", textCommentDTO);
         if (textCommentDTO.id == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
+        }
+        if (!securityService.canAccess(ctx, textCommentDTO.id, TextComment.class  )){
+            return Response.status(403, "Current user cannot access to this ressource").build();
         }
         var result = textCommentService.persistOrUpdate(textCommentDTO);
         var response = Response.ok().entity(result);
@@ -94,8 +105,11 @@ public class TextCommentResource {
     @DELETE
     @Path("/{id}")
     @RolesAllowed({AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN})
-    public Response deleteTextComment(@PathParam("id") Long id) {
+    public Response deleteTextComment(@PathParam("id") Long id, @Context SecurityContext ctx) {
         log.debug("REST request to delete TextComment : {}", id);
+        if (!securityService.canAccess(ctx, id, TextComment.class  )){
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        }
         textCommentService.delete(id);
         var response = Response.noContent();
         HeaderUtil.createEntityDeletionAlert(applicationName, true, ENTITY_NAME, id.toString()).forEach(response::header);
@@ -109,7 +123,7 @@ public class TextCommentResource {
      * @return the {@link Response} with status {@code 200 (OK)} and the list of textComments in body.
      */
     @GET
-    public Response getAllTextComments(@BeanParam PageRequestVM pageRequest, @BeanParam SortRequestVM sortRequest, @Context UriInfo uriInfo) {
+    public Response getAllTextComments(@BeanParam PageRequestVM pageRequest, @BeanParam SortRequestVM sortRequest, @Context UriInfo uriInfo, @Context SecurityContext ctx) {
         log.debug("REST request to get a page of TextComments");
         var page = pageRequest.toPage();
         var sort = sortRequest.toSort();
@@ -123,7 +137,22 @@ public class TextCommentResource {
 
 
         else{
-            result = textCommentService.findAll(page);
+
+            var userLogin = Optional
+            .ofNullable(ctx.getUserPrincipal().getName());
+        if (!userLogin.isPresent()){
+            throw new AccountResourceException("Current user login not found");
+        }
+        var user = User.findOneByLogin(userLogin.get());
+        if (!user.isPresent()) {
+            throw new AccountResourceException("User could not be found");
+        }
+            else if (user.get().authorities.size() >= 1 && user.get().authorities.stream().anyMatch(e1-> e1.equals(new Authority("ROLE_ADMIN")))){
+                result = textCommentService.findAll(page);
+
+            } else {
+                return Response.status(403, "Current user cannot access to this ressource").build();
+            }
 
         }
         var response = Response.ok().entity(result.content);
@@ -140,9 +169,12 @@ public class TextCommentResource {
      */
     @GET
     @Path("/{id}")
-
-    public Response getTextComment(@PathParam("id") Long id) {
+    @RolesAllowed({AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN})
+    public Response getTextComment(@PathParam("id") Long id, @Context SecurityContext ctx) {
         log.debug("REST request to get TextComment : {}", id);
+        if (!securityService.canAccess(ctx, id, TextComment.class  )){
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        }
         Optional<TextCommentDTO> textCommentDTO = textCommentService.findOne(id);
         return ResponseUtil.wrapOrNotFound(textCommentDTO);
     }
